\docType{methods}
\name{dbDisconnect}
\alias{dbDisconnect}
\alias{dbDisconnect,DBIDriver-method}
\alias{dbDisconnect,character-method}
\title{DBIDriver methods}
\description{
  For signature \code{DBIDriver} disconnect all
  connections. For signature \code{character}, coerce the
  character into \code{DBIDriver} and run generic.

  Character methods
}

