\docType{package}
\name{DBIextras}
\alias{DBIextras}
\alias{DBIextras-package}
\title{DBIextras}
\description{
  This adds support for common generic functions like $,
  $<-, [[, [[<-, names, and length to classes from the
  \bold{DBI} package to make interacting with databases in
  R more idiomatic.
}

